Refactor C header files
Until now, the smiol.h file defined structs and typedefs that were assumed
to be used primarily in smiol.c. However, definitions in smiol_utils.h
will need to make use of these structs and typedefs as well, with potential
dependencies as follows:

smiol.c <- smiol.h
        <- smiol_utils.h

smiol_utils.h <- smiol.h

Rather than include header files as above, adding a macro guard in smiol.h
to prevent re-definition of, e.g., structs defined in smiol.h,
the approach taken in this commit is to move the definition of all structs,
typedefs, etc. to a new file, smiol_types.h; this new file can be included
in both smiol.h and smiol_utils.h.

smiol.c <- smiol.h
        <- smiol_utils.h

smiol_utils.c <- smiol_utils.h

smiol_utils.h <- smiol_types.h

smiol.h <- smiol_types.h

Macroguards have been added to all header files to prevent re-definition
of types.
